requirements.txt
openai
pymupdf
streamlit
python-dotenv
azure-storage-blob
msal
pdf2image
pillow
pandas
opencv-python-headless
rank-bm25
azure-ai-formrecognizer
azure-ai-formrecognizer
thefuzz
pinecone
openpyxl
python-docx
